v.16
modified:   main.au3

GUI window is now created after user selects a point for the bottom of the case.
GUI window is a small window that is the color, there is text that prompts user to close window to proceed.
Text on GUI window is white or black depending on the darkness of the color.

GUI window is created after user selects a color for the pop now box itself.
GUI window is the color and text is white or black depending on the darkness of the color and it prompts use to close window to proceed.

Removed sleep code after program gets checksums of variables.