v.19
modified:   main.au3

took code from https://www.autoitscript.com/forum/topic/95920-solved-selecting-a-rectangle-to-screencapture/ to implement a drag and click method to draw a rectangle and get coords for where the boxes are located on screen.

lowered offset for caseCheckSum.

renamed variables to match rectangle selection code.

added a check to use box1 or box2 depending on which one is to the left of the case.
    - program checks the color of the left of the case to determine if the case has moved so that it can start checking the next case.

adjusted sleep times.